pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "skye.api"
authors = [{name = "Niklas Rosenstein", email = "rosensteinniklas@gmail.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
  "beartype ~=0.9.1",
  "databind.json ~=1.2.6",
  "nr.pylang.utils ~=0.1.3",
  "pyyaml ~=6.0",
]

[project.scripts]
skye-api-python = "skye.api.generator.python:main"

[project.optional-dependencies]
fastapi = [
  "fastapi ~=0.70.1",
]
dev = [
  "pytest ~=6.2.5",
  "mypy ~=0.930",
  "uvicorn ~=0.16.0",
  "types-pyyaml",
]

[project.urls]
Home = "https://gitea.com/skye/skye-api"

[tool.isort]
indent="  "
line_length=120
